+++ src/main/java/tasks/Task.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class Task {
     private String description;
     private boolean isDone;
