+++ text-ui-test/test-default/EXPECTED.TXT
